Issue #6 - Moved Project and ProjectConfig into a separate Projects folder/namespace.
