+++ OxidePatcher/NewProjectForm.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using OxidePatcher.Projects;
+using System;
 using System.Linq;
 using System.Windows.Forms;
 using System.IO;
