+++ OxidePatcher/Program.cs
@@ -1,4 +1,5 @@
 ﻿using OxidePatcher.Patching;
+using OxidePatcher.Projects;
 using System;
 using System.Globalization;
 using System.IO;
