+++ OxidePatcher/ProjectSettingsControl.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using OxidePatcher.Projects;
+using System;
 using System.IO;
 using System.Windows.Forms;
