+++ OxidePatcher/Views/ClassViewControl.resx
@@ -125,7 +125,7 @@
         AAEAAAD/////AQAAAAAAAAAMAgAAAFdTeXN0ZW0uV2luZG93cy5Gb3JtcywgVmVyc2lvbj00LjAuMC4w
         LCBDdWx0dXJlPW5ldXRyYWwsIFB1YmxpY0tleVRva2VuPWI3N2E1YzU2MTkzNGUwODkFAQAAACZTeXN0
         ZW0uV2luZG93cy5Gb3Jtcy5JbWFnZUxpc3RTdHJlYW1lcgEAAAAERGF0YQcCAgAAAAkDAAAADwMAAACC
-        LwAAAk1TRnQBSQFMAgEBFAEAATABAAEwAQABEAEAARABAAT/ARkBAAj/AUIBTQE2BwABNgMAASgDAAFA
+        LwAAAk1TRnQBSQFMAgEBFAEAATgBAAE4AQABEAEAARABAAT/ARkBAAj/AUIBTQE2BwABNgMAASgDAAFA
         AwABYAMAAQEBAAEYBgABSP8A/wD/AP8A/wD/AP8A/wD/AP8A/wD/AGkAFfYbAAP0A/YD9QPxA/UD9gP0
         HgAD8gn2A/IVAAb6MwAD9g9CA/YbAAP8A/YDZAO+A2QD9gP8GwAD8gP2CUID9gPyDAAD+gPxBvYD8QP6
         LQAD9gZCAfEB7wHwBkID9hsAA/gD8QNvA0IDbwPxA/gbAAP2BkIB8QHvAfAB8QHvAfADQgP2BgAD+gPx
